Make Future.get() cancellable.
Motivation

Currently async tasks that are waiting on NIO's EventLoopFuture will get
stuck forever if that Task is cancelled. This inhibits the good
behaviour of TaskGroup and other structured concurrency patterns that
need to have all their subtasks exit before they can.

In general users need to do more cleanup than simply cancelling a
Future, however if they forget (or don't know what that cleanup is) it's
still useful to be able to cancel an operation in this way.

Modifications

Add an extra intermediary Promise that we can fail.
Add a cancellation handler that fails that Promise.

Result

EventLoopFuture.get() immediately fails on cancellation.