Add a soudness product
Motivation:

We have a handful of scripts which are cargo-culted across the NIO
repos. This includes the soundness script for checking license headers
and unacceptable language. As regex was added to Swift 5.7 we can use it
to replace the bash in the existing script and provide a product which
can be used in the other NIO repos.

Modifications:

- Add a 'nio-soundness' product which can check for unacceptable
  language and valid license headers.
- The CLI is hand rolled to avoid bringing in a dependency on
  swift-argument-parser.
- Update the soudness script to use the new tool.
- Update the soudness script to use zsh instead of bash as the bash
  shipped on macOS is old and doesn't support globbing '**'.

Result:

Less bash, more Swift.