Move ISSUE_TEMPLATE.md to ISSUE_TEMPLATE/bug-report.md (#2273)
Motivation:

If the "New issue" button is clicked on GitHub it takes you straight to
the NIO bug report template. This makes it slightly too easy to
accidentally file a security vulnerability in public.

Modifications:

Move the ISSUE_TEMPLATE.md to ISSUE_TEMPLATE/bug-report.md, doing so
means the "New issue" will take users to a page where they can select
from filing a bug report or reporting a security vulnerability.

Result:

It's clearer when filing an issue that security vulnerabilities are
handled differently.