Rework the `NIOAsyncSequenceProducer` tests to rely less on timings
# Motivation
We see a bunch of flaky tests for the `NIOAsyncSequenceProducer` where we rely on specific timings to hit our assertions.

# Modification
This PR inverts the control of the assertions to use an event stream and await the events on the stream. This makes the tests way more reliable and gets rid of a bunch of sleeps.

# Result
Less flaky tests.